test: share paint position recorder
